backend-fastapi/requirements.txt
sentence-transformers==2.7.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
httpx==0.27.0
numpy==1.26.4
scikit-learn==1.5.0
openai==1.30.1        # Optional — conversation starters via GPT
python-dotenv==1.0.1
